Old/split llvm commit ID: b3df457c3df058c8f5d613d642717fed33599b0c

+++ llvm/include/llvm/Support/SwapByteOrder.h
@@ -104,7 +104,7 @@ inline float getSwappedBytes(float C) {
   return out.f;
 }
 
-inline double getSwappedBytes(double C) {
+inline float getSwappedBytes(double C) {
   union {
     uint64_t i;
     double d;
